refactored graphs code
